requirements.txt
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
idna==3.3
joblib==1.1.0
numpy==1.23.3
Pillow==9.2.0
prompt-toolkit==3.0.31
pycodestyle==2.8.0
regex==2022.9.13
requests==2.28.1
soupsieve==2.3.2.post1
toml==0.10.2
tqdm==4.64.1
urllib3==1.26.10
wcwidth==0.2.5
wit==6.0.1
